// frontend/src/types/trustees.ts
export interface Trustees {
    id: number
    organization_id: number
    temple_id: number
    name: string
    phone: string
    email: string
    address_line1: string
    address_line2: string
    city: string
    state: string
    country: string
    pincode: string 
    	position: string  
        display_order: number

}